requirements.txt
streamlit
gspread
pandas
plotly
google-auth